chore: Move supporting classes to `support` folder
As preparation to reorganize Visual entities classes, the first step is
to move the non-visual classes to the `support` folder.
