+++ packages/ui/src/components/Visualization/ContextToolbar/FlowType/NewFlow.tsx
@@ -3,7 +3,7 @@ import { PlusIcon } from '@patternfly/react-icons';
 import { FunctionComponent, PropsWithChildren, useCallback, useContext, useState } from 'react';
 import { useEntityContext } from '../../../../hooks/useEntityContext/useEntityContext';
 import { SourceSchemaType } from '../../../../models/camel';
-import { FlowTemplateService } from '../../../../models/visualization/flows/flow-templates-service';
+import { FlowTemplateService } from '../../../../models/visualization/flows/support/flow-templates-service';
 import { SourceCodeApiContext } from '../../../../providers';
 import { VisibleFlowsContext } from '../../../../providers/visible-flows.provider';
 import { FlowTypeSelector } from './FlowTypeSelector';
